# api/openapi.yaml
openapi: 3.1.0
info:
  title: FanBBS local vertical-slice API
  version: 0.2.0
  description: >-
    Contract for the locally runnable identity, community, private-messaging,
    notification, moderation, and administration slice. Messaging uses durable
    HTTP resources plus a polling event cursor; WebSockets and external push are
    not enabled. External SMS/email/OAuth, payments, withdrawals, lottery, and
    external object storage are also not enabled. Media uses the bounded
    local-filesystem development adapter, and SQLite is for local development
    only rather than the approved production PostgreSQL target.
servers:
  - url: http://localhost:8080/api/v1
security: []
paths:
...
  /me:
    get:
      summary: Get the authenticated private profile
      operationId: getMe
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Current user, content: {application/json: {schema: {$ref: '#/components/schemas/DataEnvelope'}}}}
    delete:
      summary: Soft-deactivate the current local account and revoke its sessions
      operationId: deactivateMe
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Account deactivated, content: {application/json: {schema: {$ref: '#/components/schemas/DataEnvelope'}}}}
  /me/account:
    delete:
      summary: Explicit alias for soft account deactivation
      operationId: deactivateMyAccount
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Account deactivated, content: {application/json: {schema: {$ref: '#/components/schemas/DataEnvelope'}}}}
  /me/profile:
    patch:
      summary: Update allowlisted profile fields
      operationId: updateMyProfile
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name: {type: string, minLength: 1, maxLength: 50}
                bio: {type: string, maxLength: 500}
                avatar_url: {type: string, format: uri, maxLength: 500}
              minProperties: 1
      responses:
        '200': {description: Updated private user, content: {application/json: {schema: {$ref: '#/components/schemas/DataEnvelope'}}}}
        '409': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
...
  /me/password:
    put:
      summary: Change the local password and revoke other sessions
      operationId: changeMyPassword
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: {type: string}
                new_password: {type: string, minLength: 8, maxLength: 128}
      responses:
        '200': {description: Password changed, content: {application/json: {schema: {$ref: '#/components/schemas/DataEnvelope'}}}}
        '401': {$ref: '#/components/responses/Error'}
...
  /me/sessions:
    get:
      summary: List active refresh-capable device sessions
      operationId: listMyDeviceSessions
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Active device sessions, content: {application/json: {schema: {$ref: '#/components/schemas/DeviceSessionListEnvelope'}}}}
        '401': {$ref: '#/components/responses/Error'}
  /me/sessions/{session_id}:
    delete:
      summary: Revoke one device session owned by the current user
      operationId: revokeMyDeviceSession
      security: [{bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/SessionID'}]
      responses:
        '200': {$ref: '#/components/responses/Interaction'}
        '401': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    SessionID: {name: session_id, in: path, required: true, schema: {type: string}}
...
  responses:
...
    Interaction:
      description: Idempotent interaction state
      content: {application/json: {schema: {$ref: '#/components/schemas/DataEnvelope'}}}
    Error:
      description: Structured API error
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
...
  schemas:
    Meta:
      type: object
      required: [request_id]
      properties: {request_id: {type: string}}
...
    DataEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {}
        meta: {$ref: '#/components/schemas/Meta'}
...
    DeviceSession:
      type: object
      required: [id, current, access_expires_at, refresh_expires_at, created_at]
      properties:
        id: {type: string}
        current: {type: boolean}
        access_expires_at: {type: string, format: date-time}
        refresh_expires_at: {type: string, format: date-time}
        created_at: {type: string, format: date-time}
    DeviceSessionListEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/DeviceSession'}}
        meta: {$ref: '#/components/schemas/Meta'}
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: {type: string}
            message: {type: string}
            field_errors: {type: object, additionalProperties: {type: array, items: {type: string}}}
            request_id: {type: string}
            retry_after: {type: integer}